requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
audioread==3.0.1
av==14.4.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
coloredlogs==15.0.1
comtypes==1.4.11
ctranslate2==4.6.0
datasets==3.6.0
decorator==5.2.1
dill==0.3.8
distro==1.9.0
faster-whisper==1.1.1
filelock==3.18.0
flatbuffers==25.2.10
frozenlist==1.7.0
fsspec==2025.3.0
greenlet==3.2.3
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.27.2
huggingface-hub==0.33.1
humanfriendly==10.0
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.26
langchain-core==0.3.71
langchain-ollama==0.3.6
langchain-text-splitters==0.3.8
langsmith==0.4.8
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
MarkupSafe==3.0.2
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.3
multiprocess==0.70.16
networkx==3.5
numba==0.61.2
numpy==2.2.6
ollama==0.5.1
onnxruntime==1.22.0
openai==1.93.0
openai-whisper==20250625
opencv-python==4.12.0.88
orjson==3.11.0
packaging==25.0
pandas==2.3.0
pinecone-client==6.0.0
pinecone-plugin-interface==0.0.7
platformdirs==4.3.8
pooch==1.8.2
propcache==0.3.2
protobuf==6.31.1
pvporcupine==3.0.5
pyarrow==20.0.0
PyAudio==0.2.14
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pypiwin32==223
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
pyttsx3==2.99
pytz==2025.2
pywin32==311
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.16.0
semantic-version==2.10.0
sentencepiece==0.2.0
setuptools-rust==1.11.1
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soxr==0.5.0.post1
SpeechRecognition==3.14.3
SQLAlchemy==2.0.41
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.2
torch==2.7.1
tqdm==4.67.1
transformers==4.53.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
xxhash==3.5.0
yarl==1.20.1
zstandard==0.23.0